:recycle: Move create_release_tag into repo::version_tag
