+++ src/lib.rs
@@ -7,5 +7,6 @@ pub mod test_util;
 pub use crate::changes::Changes;
 pub use crate::changes::SemanticVersionAction;
 pub use crate::config::{CommitParserOptions, SemanticReleaseConfig};
+pub use crate::repo::prelude::create_release_tag;
 pub use crate::repo::prelude::render_tag;
 pub use crate::repo::prelude::RepositoryExtension;
